requirements.txt
pytest
pytest-asyncio
pytest-xdist
httpx
allure-pytest
pydantic